move useEffect from profile_details index to hooks
