+++ src/screens/profile_details/index.tsx
@@ -1,5 +1,4 @@
-import React, { useEffect } from 'react';
-import { useRouter } from 'next/router';
+import React from 'react';
 import { NextSeo } from 'next-seo';
 import useTranslation from 'next-translate/useTranslation';
 import {
@@ -7,29 +6,16 @@ import {
   LoadAndExist,
   DesmosProfile,
 } from '@components';
-import { chainConfig } from '@configs';
 import { useStyles } from './styles';
 import { Connections } from './components';
 import { useProfileDetails } from './hooks';
 
 const ProfileDetails = () => {
-  const router = useRouter();
   const { t } = useTranslation('profiles');
   const classes = useStyles();
   const {
     state,
   } = useProfileDetails();
-
-  useEffect(() => {
-    const regex = /^@/;
-    const profileDtag = router.query.dtag as string;
-    const regexCheck = regex.test(profileDtag);
-    const configProfile = chainConfig.extra.profile;
-
-    if (!regexCheck || !configProfile) {
-      router.replace('/');
-    }
-  }, []);
 
   return (
     <>
